@@ -14,7 +14,7 @@
   },
   "scripts": {
     "build": "rm -rf lib/* && tsc",
-    "lint": "tslint -p tsconfig.json -c tslint.json",
+    "lint": "eslint . --ext .js,.jsx,.ts,.tsx",
     "lint:types": "tsc --noEmit",
     "format": "prettier --write ./src/**/*.ts",
     "format:check": "prettier --list-different ./src/**/*.ts",
@@ -33,14 +33,16 @@
     "test": "test"
   },
   "devDependencies": {
-    "@types/jest": "^25.1.4",
-    "@types/node": "^13.9.5",
-    "jest": "^25.2.4",
-    "prettier": "^2.0.2",
-    "semantic-release": "^15.14.0",
-    "ts-jest": "^25.2.1",
-    "tslint": "^6.1.0",
-    "typescript": "^3.8.3"
+    "@types/jest": "^26.0.14",
+    "@types/node": "^14.11.2",
+    "@typescript-eslint/eslint-plugin": "^4.2.0",
+    "@typescript-eslint/parser": "^4.2.0",
+    "eslint": "^7.9.0",
+    "jest": "^26.4.2",
+    "prettier": "^2.1.2",
+    "semantic-release": "^17.1.2",
+    "ts-jest": "^26.4.0",
+    "typescript": "^4.0.3"
   },
   "keywords": [
     "encoding",
