@@ -1,6 +1,5 @@
 import { Context, Recogniser } from '.';
-
-var match = require('../match').default;
+import match, { Match } from '../match';
 
 /**
  * This is a superclass for the individual detectors for
@@ -15,7 +14,7 @@ class ISO_2022 implements Recogniser {
     return 'ISO_2022';
   }
 
-  match(det: Context) {
+  match(det: Context): Match | null {
     /**
      * Matching function shared among the 2022 detectors JP, CN and KR
      * Counts up the number of legal an unrecognized escape sequences in
@@ -29,16 +28,16 @@ class ISO_2022 implements Recogniser {
      * @return match quality, in the range of 0-100.
      */
 
-    var i, j;
-    var escN;
-    var hits = 0;
-    var misses = 0;
-    var shifts = 0;
-    var quality;
+    let i, j;
+    let escN;
+    let hits = 0;
+    let misses = 0;
+    let shifts = 0;
+    let quality;
 
     // TODO: refactor me
-    var text = det.fInputBytes;
-    var textLen = det.fInputLen;
+    const text = det.fInputBytes;
+    const textLen = det.fInputLen;
 
     scanInput: for (i = 0; i < textLen; i++) {
       if (text[i] == 0x1b) {
@@ -47,7 +46,7 @@ class ISO_2022 implements Recogniser {
           escN < this.escapeSequences.length;
           escN++
         ) {
-          var seq = this.escapeSequences[escN];
+          const seq = this.escapeSequences[escN];
 
           if (textLen - i < seq.length) continue checkEscapes;
 
