@@ -1,12 +1,12 @@
+import { Context, Recogniser } from "./encoding";
 
 export interface Match {
   confidence: number;
   name: string;
-  lang: string;
+  lang?: string;
 }
 
-// @ts-ignore
-export default (det, rec, confidence, name, lang): Match => ({
+export default (det: Context, rec: Recogniser, confidence: number, name?: string, lang?: string): Match => ({
   confidence,
   name: name || rec.name(det),
   lang,
